ml-api/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
Pillow==11.1.0
ultralytics==8.2.0
easyocr==1.7.1